add base class for Setup
